Update README.md
Not sure if I did this right, so look at it closely. Bill copies all three lines  on the **Install the Package and Dependencies** at the same time when he was first trying and it was my opinion that they should be separate. AND I wanted to try a commit